// assets/games/tongits/script/proto/tongits.ts
export interface DrawCardRes {
  /** 抽到的牌，只有自己知道 */
  drawnCard: number;
  /** 是否有Tongits */
  hasTongits: boolean;
  /** 手牌数量 (对其他玩家可见) */
  handCardCount: number;
  /** 自己的手牌  自己会有这个值 */
  groupCards: Cards[];
}

/**
 * 抽牌广播
 * MessageType: TONGITS_DRAW_BROADCAST (3021)
 */
export interface DrawCardBroadcast {
  playerId: number;
  userId: number;
  /** 当自动抽牌时  自己会有这个值 */
  drawnCard: number;
  /** 手牌数量 (对其他玩家可见) */
  handCardCount: number;
  /** 自己的手牌  自己会有这个值 */
  groupCards: Cards[];
}

/**
 * 出牌
 * MessageType: TONGITS_MELD_REQ (3005)
 */
export interface MeldCardReq {
  cards: number[];
}
// ...
export interface LayOffCardReq {
  /** 要补的牌 */
  card: number;
  /** 目标牌组的拥有者ID */
  targetPlayerId: number;
  /** 目标牌组的ID */
  targetMeldId: number;
}
// ...
export interface DiscardCardReq {
  card: number;
}
// ...
export interface TakeCardReq {
  /** 手中用来组合的牌 */
  cardsFromHand: number[];
}
// ...
export interface ChallengeReq {
  /** 2:发起 3:接受 4:拒绝 */
  changeStatus: number;
}
// ...
export interface GameResultDetailsRes {
  winnerId: number;
  playerResults: PlayerResult[];
}
// ...
export interface SwitchAutoGroupCardsReq {
  isAuto: boolean;
}

/**
 * 开启/关闭自动组牌
 * MessageType: TONGITS_SWITCH_AUTO_GROUP_CARDS_RES (3037)
 */
export interface SwitchAutoGroupCardsRes {
  isAuto: boolean;
  /** 自己的手牌 */
  groupCards: Cards[];
}

/**
 * 玩家手动组牌
 * MessageType: TONGITS_GAME_PLAYER_GROUP_CARDS_REQ (3038)
 */
export interface GamePlayerGroupCardsReq {
  /** 需要变化的牌组列表  组成新的group 不需要传参 */
  targetGroupCards: Cards[];
}
// ...
export const MeldCardReq = {
  encode(message: MeldCardReq, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    writer.uint32(10).fork();
    for (const v of message.cards) {
      writer.int32(v);
    }
    writer.ldelim();
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): MeldCardReq {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMeldCardReq();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag === 8) {
            message.cards.push(reader.int32());

            continue;
          }

          if (tag === 10) {
            const end2 = reader.uint32() + reader.pos;
            while (reader.pos < end2) {
              message.cards.push(reader.int32());
            }

            continue;
          }

          break;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): MeldCardReq {
    return { cards: globalThis.Array.isArray(object?.cards) ? object.cards.map((e: any) => globalThis.Number(e)) : [] };
  },

  toJSON(message: MeldCardReq): unknown {
    const obj: any = {};
    if (message.cards?.length) {
      obj.cards = message.cards.map((e) => Math.round(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<MeldCardReq>, I>>(base?: I): MeldCardReq {
    return MeldCardReq.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<MeldCardReq>, I>>(object: I): MeldCardReq {
    const message = createBaseMeldCardReq();
    message.cards = object.cards?.map((e) => e) || [];
    return message;
  },
};
// ...
export const LayOffCardReq = {
  encode(message: LayOffCardReq, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.card !== 0) {
      writer.uint32(8).int32(message.card);
    }
    if (message.targetPlayerId !== 0) {
      writer.uint32(16).int64(message.targetPlayerId);
    }
    if (message.targetMeldId !== 0) {
      writer.uint32(24).int32(message.targetMeldId);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): LayOffCardReq {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseLayOffCardReq();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.card = reader.int32();
          continue;
        case 2:
          if (tag !== 16) {
            break;
          }

          message.targetPlayerId = longToNumber(reader.int64() as Long);
          continue;
        case 3:
          if (tag !== 24) {
            break;
          }

          message.targetMeldId = reader.int32();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): LayOffCardReq {
    return {
      card: isSet(object.card) ? globalThis.Number(object.card) : 0,
      targetPlayerId: isSet(object.targetPlayerId) ? globalThis.Number(object.targetPlayerId) : 0,
      targetMeldId: isSet(object.targetMeldId) ? globalThis.Number(object.targetMeldId) : 0,
    };
  },

  toJSON(message: LayOffCardReq): unknown {
    const obj: any = {};
    if (message.card !== 0) {
      obj.card = Math.round(message.card);
    }
    if (message.targetPlayerId !== 0) {
      obj.targetPlayerId = Math.round(message.targetPlayerId);
    }
    if (message.targetMeldId !== 0) {
      obj.targetMeldId = Math.round(message.targetMeldId);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<LayOffCardReq>, I>>(base?: I): LayOffCardReq {
    return LayOffCardReq.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<LayOffCardReq>, I>>(object: I): LayOffCardReq {
    const message = createBaseLayOffCardReq();
    message.card = object.card ?? 0;
    message.targetPlayerId = object.targetPlayerId ?? 0;
    message.targetMeldId = object.targetMeldId ?? 0;
    return message;
  },
};
// ...
export const DiscardCardReq = {
  encode(message: DiscardCardReq, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.card !== 0) {
      writer.uint32(8).int32(message.card);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): DiscardCardReq {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseDiscardCardReq();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.card = reader.int32();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): DiscardCardReq {
    return { card: isSet(object.card) ? globalThis.Number(object.card) : 0 };
  },

  toJSON(message: DiscardCardReq): unknown {
    const obj: any = {};
    if (message.card !== 0) {
      obj.card = Math.round(message.card);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<DiscardCardReq>, I>>(base?: I): DiscardCardReq {
    return DiscardCardReq.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<DiscardCardReq>, I>>(object: I): DiscardCardReq {
    const message = createBaseDiscardCardReq();
    message.card = object.card ?? 0;
    return message;
  },
};
// ...
export const TakeCardReq = {
  encode(message: TakeCardReq, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    writer.uint32(10).fork();
    for (const v of message.cardsFromHand) {
      writer.int32(v);
    }
    writer.ldelim();
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): TakeCardReq {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTakeCardReq();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag === 8) {
            message.cardsFromHand.push(reader.int32());

            continue;
          }

          if (tag === 10) {
            const end2 = reader.uint32() + reader.pos;
            while (reader.pos < end2) {
              message.cardsFromHand.push(reader.int32());
            }

            continue;
          }

          break;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): TakeCardReq {
    return {
      cardsFromHand: globalThis.Array.isArray(object?.cardsFromHand)
        ? object.cardsFromHand.map((e: any) => globalThis.Number(e))
        : [],
    };
  },

  toJSON(message: TakeCardReq): unknown {
    const obj: any = {};
    if (message.cardsFromHand?.length) {
      obj.cardsFromHand = message.cardsFromHand.map((e) => Math.round(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<TakeCardReq>, I>>(base?: I): TakeCardReq {
    return TakeCardReq.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<TakeCardReq>, I>>(object: I): TakeCardReq {
    const message = createBaseTakeCardReq();
    message.cardsFromHand = object.cardsFromHand?.map((e) => e) || [];
    return message;
  },
};
// ...
export const ChallengeReq = {
  encode(message: ChallengeReq, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.changeStatus !== 0) {
      writer.uint32(8).int32(message.changeStatus);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): ChallengeReq {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseChallengeReq();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.changeStatus = reader.int32();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ChallengeReq {
    return { changeStatus: isSet(object.changeStatus) ? globalThis.Number(object.changeStatus) : 0 };
  },

  toJSON(message: ChallengeReq): unknown {
    const obj: any = {};
    if (message.changeStatus !== 0) {
      obj.changeStatus = Math.round(message.changeStatus);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ChallengeReq>, I>>(base?: I): ChallengeReq {
    return ChallengeReq.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ChallengeReq>, I>>(object: I): ChallengeReq {
    const message = createBaseChallengeReq();
    message.changeStatus = object.changeStatus ?? 0;
    return message;
  },
};
// ...
export const SwitchAutoGroupCardsReq = {
  encode(message: SwitchAutoGroupCardsReq, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.isAuto !== false) {
      writer.uint32(8).bool(message.isAuto);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): SwitchAutoGroupCardsReq {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSwitchAutoGroupCardsReq();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.isAuto = reader.bool();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SwitchAutoGroupCardsReq {
    return { isAuto: isSet(object.isAuto) ? globalThis.Boolean(object.isAuto) : false };
  },

  toJSON(message: SwitchAutoGroupCardsReq): unknown {
    const obj: any = {};
    if (message.isAuto !== false) {
      obj.isAuto = message.isAuto;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SwitchAutoGroupCardsReq>, I>>(base?: I): SwitchAutoGroupCardsReq {
    return SwitchAutoGroupCardsReq.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SwitchAutoGroupCardsReq>, I>>(object: I): SwitchAutoGroupCardsReq {
    const message = createBaseSwitchAutoGroupCardsReq();
    message.isAuto = object.isAuto ?? false;
    return message;
  },
};